Revert "Skip haproxy with setup-infrastructure for upgrades"
This reverts commit befd8424e2efd4e1bebe89b5085032bf120de148.

Reason for revert:

This change is not needed anymore for A release, especially with
separated-haproxy-service-config feature in place.

It creates a problem now, because haproxy-install.yml is responsible for
creating "base" haproxy service which is required for horizon to work.
If horizon does not work during setup-openstack.yml, tempest tests
may fail(along with the whole upgrade job in our CI)

That's why we need to run haproxy-install.yml, traditionally during
setup-infrastructure.

Change-Id: I5531e56610a8a63f601867d236831ce7d5e4d583
Needed-By: https://review.opendev.org/c/openstack/openstack-ansible-os_horizon/+/879517