Stop gathering local software_versions for services
With latest ansible-core playbooks started failing on adding extra compute
or controller nodes, when cinder/nova playbooks run with limits.

This happens as we're trying to reply on local facts for hosts that are
expired. At the same time, it's not always possible to collect them, as some computes
can be down while adding another one.

With that we're simplifying flow and avoid old process of
restarting services or executing migrations based on local facts.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-os_nova/+/880147
Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-os_cinder/+/880210
Closes-Bug: #2009834
Change-Id: I44dc8567e9a93f91327202de1bf88a067266711d