Fix haproxy_nova_api_metadata_service config
In [1] we merged an incorrect haproxy config for
haproxy_nova_api_metadata_service[2].

First of all, `haproxy_ssl_all_vips` is not used because `haproxy_bind`
is defined, so this service has only one frontend(internal vip).
It can be dropped.

Only `haproxy_ssl` has an effect, but because it's an internal ip
address, its value should be set to `"{{ haproxy_ssl_all_vips }}"`
which is generally used to define protocol for internal VIPs(yes, the
name may be a bit misleading). We did the same for repo_server in [3].

Currently, nova metadata service listens on https and neutron tries to
reach it over http because `openstack_service_internaluri_proto` is set
to 'http' by default. As the result, neutron is not able to reach
metadata service.

[1] https://review.opendev.org/c/openstack/openstack-ansible/+/866119
[2] https://opendev.org/openstack/openstack-ansible/blame/commit/dc1f76c82358e5efd55a3a5ee24b2c99b479a165/inventory/group_vars/haproxy/haproxy.yml#L366-L367
[3] https://review.opendev.org/c/openstack/openstack-ansible/+/876426

Change-Id: I54a78ad3fc35775a104732a544c5e1ff9c14827d