Gather generic masakari facts
With commit [1] we moved extra facts gathering to pre_tasks
but with that we did not enable generic facts gathering, which
led to regression. So we cover this by ensuring that generic
facts are also gathered and not only extra ones.

[1] https://opendev.org/openstack/openstack-ansible/commit/8bc9b167ab9a4854aab5d49a10709c27e96ff833
Closes-Bug: #1979145

Change-Id: Ibb9dc3377a4de06af25281bf777b16faad16d261
(cherry picked from commit bb4f1c7b2a46b30de8bca3e695938022efe6be89)