Fix blazar haproxy service
In [1], missing blazar haproxy server was added.
Turns out that blazar does not support '/healthcheck' and all API
requests require authentication.
It makes it hard for haproxy to monitor backends because they respond
with 401 response code(unauthorized).
With this patch, haproxy will accept 401 response code along with 200
for blazar service.

[1] https://review.opendev.org/c/openstack/openstack-ansible/+/880564

Change-Id: I882e9257b41681a25bda5e200df7e8bd8700c186