Revert "Add magnum tempest URL"
This reverts commit 03353d998a2cc0fe8fdac358c16e70a47f1ff029.

Reason for revert:
Hardcoded protocol is problematic because from now on, backends may be
listening on TLS.
In [1] there was an attempt to leverage
`openstack_service_internaluri_proto` variable.
Unfortunately due to issues with config_template[2] we decided to
abandon it and avoid defining magnum_url(at least for now).


[1] https://review.opendev.org/c/openstack/openstack-ansible/+/763049
[2] https://bugs.launchpad.net/openstack-ansible/+bug/1649381

Change-Id: I2bfc271a52982577b94743e989a69d4a31fcd2d6