Bump down etcd version for zun
Docker libkv [1] does not support etcd of version >=3.4,
as otherwise ErrInvalidJSON exception is raised. Since
docker cluster-store is deprecated and removed in 23.0.0,
support of etcd won't be fixed.

At the same time while zookeeper is also a supported driver,
it's usage causes docker to crash on attempt to use it.

So bumping etcd version down is the only solution at the
moment.

[1] https://github.com/docker/libkv

Needed-By: https://review.opendev.org/c/openstack/openstack-ansible-os_zun/+/880288
Change-Id: I9a13d836f7f266c54283125dc09d11feeea23d3e