Force enable openstack_service_token_roles_required
To reduce downtime during major upgrades, we've disabled
requirement of providing service tokens along with user ones.

However with OSSA-2023-003 enabling service tokens is a requirement. So
we have no other choice but to change the value and infrom users about
changes with release note.

Distro jobs are removed to prevent circular dependency

Needed-By: https://review.opendev.org/c/openstack/openstack-ansible-os_cinder/+/884211
Needed-By: https://review.opendev.org/c/openstack/openstack-ansible-os_nova/+/884213
Change-Id: I1fd9c7fe96dbba708e3fba2f769838605bf9ad05