Pin version of setuptools
We used to pin version of setuptools in global-requirements-pins
until it got defined in upper-constraints. Since setuptools no longer
kept in u-c since Zed, it makes sense to return pinning back.

On top of that a major setuptools version was released recently with
breaking changes, that affect gnocchi at least [1]

[1] https://github.com/gnocchixyz/gnocchi/issues/1304

Change-Id: I5932e8b66bb4674fb1b39442fea467c075b30989