Use include_role in task to avoid lack of access to vars
This patch updates the security hardening playbook to use include_role
within a task versus using 'roles' directly to fix cases where
apply_security_hardening is set to False. Some change to Ansible
appears to limit access to vars when the role is skipped, resulting
in failures. The side effect of this change is the role is skipped
entirely (when applicable) versus the individual tasks being skipped,
which speeds up deployment times.

Change-Id: Iaf81f778ffd9fcf6572e03e8ce73ffef46bd98aa