Return PIP_OPTS for load_nodepool_pip_opts
At the moment for some scenarios, like linters, where ansible bootstrap
is skipped, PIP_OPTS are undefined and linters fail due to that.

With patch we define PIP_OPTS to an empty value if
we're not in CI.

Alternatively we can patch scenarios independently
not to rely on existance of PIP_OPTS or export default in
gate-check-commit instead.

Change-Id: If451b4bc12bb0b0bfe3fe20494e99a07b59ab798