haproxy: fix health checks for serialconsole in http mode
The serialconsole proxy appears to be a very simple WebSocket
proxy with no obvious support for many HTTP methods. As such
a HEAD or GET to / returns a 405.

This patch adjusts the expected status code to ensure the service
isn't marked as down, but it is worth noting that the health
checks do make the serialproxy service logs a little noisy.

Change-Id: Ie2e06c99b9c99435e29a321b373d6a937f4123f4