Avoid breaking horizon during Antelope upgrade
During Antelope upgrade, haproxy-install.yml configures base service
listening on ports 80 and 443.
But at this point, horizon frontend is not removed yet so haproxy has
two frontends listening on 80/443.
During os-horizon-install.yml, horizon frontend is removed and
appropriate haproxy maps config is applied.
As a result, between haproxy-install.yml and os-horizon-install.yml,
horizon will encounter intermittent issues.
This patch triggers horizon service reconfiguration during Antelope
upgrade when running haproxy-install.yml playbook.
To be precise: when horizon is already deployed in an environement but
its maps config is not there yet, haproxy will trigger reconfiguration
of horizon service.
This change is intendent to land only in 2023.1. Upgrades from 2023.1
to 2023.2 or 2024.1 are not affected.

Closes-Bug: #2028944
Change-Id: I9df78f286c1fcc3204b35928829a981addf876e8