Respect haproxy_bind_internal_lb_vip_* variables
At the moment we assume that haproxy should be fine listening on
internal_lb_vip_address, but in real life deployments these are FQDN
and in case of using DNS RR, this assumption is invalid.

We can be smarter and check if haproxy_bind_internal_lb_vip_* variables
are defined, and fallback to previous behaviour if not.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-haproxy_server/+/852039
Change-Id: Ic0b9646d566425878930eb88745e35f9e6cc2e11