Apply rate limit for journald in AIO builds
At the moment we experience DISK_FULL issue in CI for upgrade
jobs on Rocky. This happens as our log files are more then
available per-job limit, which is capped at 5Gb. In order
to see what fills up the diskspace and reduce overall disk consumption,
we apply rate limiting and define maximum size of files as well as
total diskspace that can be consumed by logs not to exceed that limit.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-openstack_hosts/+/892151
Change-Id: I960a4377f6cfc6977a4d0df0e53744cdfb539b1e