Add default name for user collections file
The get-ansible-collection-requirements playbook cannot be used
standalone without passing an extra var for the name of the
user collections file.

This patch allows the playbook to be run standalone and the
default name will be used automatically, still allowing an override
to be used as currently happens in scripts/bootstrap-ansible.sh

Sonetimes it is desireable to have a simple way to update the
installed collections without running other parts of
bootstrap-ansible, or look up flags to disable other side-effects.

Change-Id: I27131ecaa3bc0ef459d2c353be9531608dd9d9ab