Ensure tempest include and exclude lists all use unique names
When deploying a complex AIO with several services enabled, currently
duplicate definitions are made for the tempest test include and
exclude lists. It is down to ansible variable precedence to
determine which one is actually used.

This patch uses the changes made in [1] to allow tempest include and
exlude lists to be defined in many variables and aggregated at the
point the config files for tempest are written.

[1] https://review.opendev.org/c/openstack/openstack-ansible-os_tempest/+/891579

Change-Id: If047578c283e85b38e73a5f117289f42150f3daf