Gather extra networking facts for keepalived
By default keepalived pingable check tries to verify the gateway which
in it's turn requires to have appropriate facts defined.

At the moment re-running haproxy playbook as a standalone instance
when facts are outdated will lead to re-configuration of keepalived
to check 127.0.0.1 as a track script.

So we gather extra network facts to identify the gateway for
pingable script.

Change-Id: I4e72e81d05d8d70e76730a46687dd06f3e66ebc6