Map default value of rabbitmq_management_ssl to haproxy_ssl
At the moment all haproxy backends are defined if TLS should be used by
using `haproxy_ssl` variable. If deployer don't want to have SSL, they
are supposed to use the variable for that. However, the only service that
is not respecting that is RabbitMQ management interface.

As a result haproxy fails with the invalid configuration, since
certificates are not provisioned when `haproxy_ssl` is False.
So configuration at the end is invalid as reffer to the certificate
that does not exist on the host and was not even issued.

Change-Id: Idc924d4ee485c8e6efc15b90df90ba5021a106e4