abstract bootstrap host disk partition names
The code in this playbook assumes that the path to a partition is just
the partition number appended to the disk path. However when working
with NVMe disks, the partition will actually be "pN" where N is the
partition number and p is the literal character. To attempt to be
agnostic use lsblk to show all the partitions on the disk, since it was
just reformatted and only has the partitions this playbook creates. This
is then stored in a new temporary fact that can be used throughout the
rest of the playbook.

Change-Id: I661696c275fbc2e91d049303f5498ff8d401fc83