Allow installing collections from repos containing more than one
A git repo may contain one collection, several collections at
the top level or be organised into namespace/collection named
directories. See [1].

The '#' symbol is used to identify a particular collection from
a repository to install. This patch ensures that the paths
searched in CI for zuul source directories do not include parameters
beyond the '#' symbol, and that the contents of the collection
requirements file is also correct for installing any required
collections using the 'dir' type.

[1] https://docs.ansible.com/ansible/latest/dev_guide/developing_collections_distributing.html#specifying-the-collection-location-within-the-git-repository

Change-Id: I081212819c61b46c2ad143025e7287cd3ce5390a