Skip installing curl for EL
Due to the bug in Ansible [1] it does not resolve requirement of /usr/bin/curl
properly, always trying to install `curl-minimal`, which would fail on systems with
already installed `curl`.

Since `curl` presence is essential for DNF, we should be able to skip installation
of curl to EL systems, as they are all shipped with it as of today.

[1] https://github.com/ansible/ansible/issues/82461

Closes-Bug: #2046172
Change-Id: I65d3fa2e3d6d38f0a51ce1f402ac25d976b67b89
(cherry picked from commit 60a1196fe69730d9c2aed51b72a8495464e6c5ed)