Fix shallow_since for rabbitmq_server on 2023.1
It appears that the requirements update in
https://review.opendev.org/c/openstack/openstack-ansible/+/904303
might have set shallow_since incorrectly. As far as I can see this
is generally set to the day before the referenced commit. In this
case it was set to the day after.

This is causing upgrade job CI failures now that the referenced
commit isn't at the top of stable/2023.1

Change-Id: I0da61d324a09d8304a7e3371df5f21dc4774ab1a
Needed-By: https://review.opendev.org/c/openstack/openstack-ansible-os_magnum/+/904508