Treat dashes/underscores as interchangeable symbols for container names
This patch aims to treat dashes and underscores equally in the existing
inventory.

In case a host already exists in the inventory, but has different separators,
treat underscore and dash symbol as interchangeable.

This is needed for futher coversion of container names to spearated with only dashes,
while not adding new records to the inventory

Change-Id: I63cd14c3353d7a9d7ea4d96155be26697ee4fa40