[doc] Remove guidance to drain RMQ which can result in failures
oslo_messaging isn't set up to handle this situation and appears
to expect the node to come back (as would happen during a version
upgrade). As a result, client processes don't attempt to re-create
queues which lived on the node being upgraded. This situation
doesn't auto-recover which can be very disruptive.

Change-Id: I608202d665417682ce7309bb9bf4b52a2cf4373e