Enable RabbitMQ Quorum Queues by default
This patch finalizes [1] and [2] topics and enables usage of quorum
queues by default along with adding release note and documentation
around the change.

We also convert previous quorum job to the classic HA queues
test, not to break an upgrade path for those who want to switch
to Quorum after OpenStack upgrade.

[1] https://review.opendev.org/q/topic:%22osa/messaging_improvements%22
[2] https://review.opendev.org/q/topic:%22osa/rmq-migrate%22

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible/+/919902
Change-Id: I80e24343f28f712492d378186d0a0fa5832907e2