Enable RabbitMQ feature flags pre-upgrade
This task must be carried out before any RabbitMQ instances are
stopped, so must live in the playbook as opposed to the role.

This complements the post-upgrade enablement which happens within
the role itself, and covers the case where older versions of the
role have not already enabled the relevant flags.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-rabbitmq_server/+/919701
Change-Id: Idf33b6810e1afb3a89dad5e65a48e86ac0a58a67