Remove old horizon base mapping
In case of environment upgrades, with they want to enable Skyline
dashboard after upgrade, we need to ensure that old horizon mapping
with higher priority is removed first.
Otherwise it will lead to incorrect behaviour of Skyline/Horizon.

I do not push this to master, as this is only relevant for upgrades to
2024.1. Upgrade to 2024.2 is supported only from 2024.1, where it will
already be covered.

Change-Id: I13c4d874b6c99732758628313188b60ad7d85c4d