Delete glusterd data along with container data
When a user deletes a container and selects
`force_containers_data_destroy` it makes sense to also delete gluster
data for the container as well.

Otherwise, once containers respawn, they already do have some stale
data from the previous container which may not really work out and
would require manual cleanup.

With that, when user selects to delete all container data, it's expected
that all data for container should be destryed, so that it would be
created from scratch.

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-openstack_hosts/+/924994
Change-Id: I85e3768b9611e498cbfefd1e02c12c8b94a6f30b