Properly apply `always` tag to install_defaults
In order to complete a playbook successfully we rely on having
deploy-type specific variables, which are provided by install_defaults
role.

This used to be a simple include_vars statement instead, which was
always executed with any tags, but it's not a case now with inlcude_role
where we should explicitly take care about always running the role
regardless of tags provided.

Change-Id: I7a7112dd6a7c3a360dc6c0d206a855c220a027a6