Ensure "parallel" package is installed for CI log collection
The log collection requirements are installed at the start of the
job, running from the starting upgrade branch. We need to ensure
that the tools needed by the destination branch are present.

In the case of a 2023.1 to 2024.1 SLURP upgrade, these tools are
missing so this patch is a partial backport of [1].

[1] Ief03c33b76eac6e256477cfaa16c8a59acd7d702

Change-Id: Ic3b4b543e58904f8f6fef36e7ae7cb97638c1d3c
(cherry picked from commit 4454d1a15b17dcb052c7489c2db54d8c69078b10)