Add centralized defenition for molecule jobs
In order to centrally manage jobs launching molecule a defined jobs and
project-template added.

At the moment it's relying on tox to execute molecule. While we can
run molecule with gate-check-commit.sh, it's obviously better for
development to have at least some local testing
present, that does not heavily depend on in-zuul logic.

Depends-On: https://review.opendev.org/c/openstack/requirements/+/938224
Change-Id: Ic9b3fa32236279c1cf991593e831bd950f1adc0a