Add noop molecule job to the integrated repo
In order to verify that molecule requirements are satisfied we add a
noop molecule scenario that installs requirements and run a noop
task.

We can expand the scenario in the futute or add new ones, though for
now goal is to ensure that updating requirements is not breaking
other roles having valid scenarios around.

Change-Id: Id0abc217d404af4a08532707e60d738613c9a7ed