Add additional commented RabbitMQ policy to manage segment sizes
The default RabbitMQ segment sizes of 500MB mean you may need to
acquire many days or weeks worth of messages in each fanout/stream
queue before the x-max-age setting takes effect. Subject to the
number of streams in use across the deployment, this could consume
many tens of GB of disk space.

This example shows an additional policy which has been tested in
order to combat this in the absence of suitable per-stream settings
in oslo.messaging. The downside is that this must be applied before
any services create their streams.

Change-Id: I69699f99a88b0ba8c8ee283ecbd3b4ed0c98a373