Move argument parsing for dynamic_inventory to generate
At the moment osa_toolkit is installable as a package. With recent
changes to PBR for python3.12 console_scripts are no longer
installed "as is", but only a simple import is created. This change
results in broken `openstack-ansible-inventory` binary, as `inventory`
can not be imported.

To avoid the issue and follow our existing logic for rest of tooling,
we move argument parsing logic to generate.py itself, allowing it to
be self-sufficient and be leaving a dynamic_inventory.py as a simple
pointer which needs to stay in inventory folder for gropup_vars to
load properly.

Change-Id: Ie6d68f9c7b91d88736c5fc67c997cffd812afc61