Prevent inventory-manger failing with arbitrary groups
In case an arbitrary record is added to openstack_user_config
which ends with `_hosts` where metal setup is used, it can result in
inventory-mange.py crash as the group will not be mapped.

This case is handled now and script will properly display group
assigment for this scenario.

Related-Bug: #2102132
Change-Id: I73daae0334876399cb429b5a88376474abfef79e