Use first utility host as default host for openstack_resources
At the moment openstack_service_setup_host is not actually respected as
group_vars defaults provisionment happens during runtime of
the playbook.

With that, openstack_service_setup_host does have the same value
by default for both distro and source installations, which is the
first utility host.

Let's use this value as default since it might fit most of the usecases.

Closes-Bug: #2096939
Change-Id: Id3eb93a3c226185c614e324609bf7be5d789f248