Fix role and collection dependency
With merge of [1] Depends-On for roles and services got broken as'
ZUUL_SRC_PATH is not really defined in the role, but lookup returnes
an empty string instead of falling back to the default value.

Also default value was pointing to the wrong directory anyway,
and not to /home/zuul.

Adjustments here should be fixing the behaviour.

[1] https://review.opendev.org/c/openstack/openstack-ansible/+/939151/27

Change-Id: I5eeb808b8eb2c21ef2e3ca898e872098152d0f1f