Remove LIBSYSTEMD_VERSION override for AIO
This override for CI was introduce for CentOS 8.3 in order to
supply a valid version of systemd for wheels build of systemd-python.

Suc override seems no longer needed, but also causing issues
in AIO for EL10, where it defines LIBSYSTEMD_VERSION=257.5,
while an integer value is expected by the builder.

[1] https://review.opendev.org/q/I94008de1fe97400ab37de39d03599fa9f53ab3f0

Change-Id: I1c2ea26bb585b869629b2d1be29d30ed3765f3bf