Make AIO networks configurable
We carry *_range variables without any real use as of today. With
that sometimes it makes total sense to override some AIO subnet
range in case it does intersect with already existing subnet on AIO VM.

This also should be useful and more flexible for future patches
which may add situational complexity.

Change-Id: Ibe46a54a4eca981a73a9b16e0945ec2944c1db87
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>