[aio] Include loopback preparations conditionally
Some niche scnarios, like Swift or Manila are executed for each
scenario, which is not necessary and might be confusing or
overwhelming for operators.

Let's limit amount of prepared loopback only to used scenarios.

Variable `bootstrap_host_loopback_machines` is not used anywhere in
AIO bootstrap logic, so can be safely removed.

Change-Id: Ic52c923f72bedaf6863bc55e4a43f0f541d9b5b2
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>