Do not disable configure_mirrors extra repos for debian
This does not work with the current zuul images as there is a backports
repo config present in the built image that is not replaced by the
CI region specific config when the extra repos are disabled.

As the currently built images specify the '-int' (internal) interface
for the mirror server, this leads to job failures as the internal
mirror url from the image is not accessible when the image is booted
in a different region to that which it was built in.

Change-Id: I4b387fc25b56d70ee8a36d1fe91b3cd6933e2dc4
Signed-off-by: Jonathan Rosser <jonathan.rosser@rd.bbc.co.uk>