Serve proxy on a standalone interface
Our proxy scenario is a bit unique, as not only we want to test
that hosts access Squid and can perform operations through it,
but also that they can work without access to Public VIP (through
internal VIP only).

With that access to proxy is required from the very beginning of
deployment, so we can not offload interface creation to openstack_hosts.

So a separate dummy interface is added which is used only to serve as a
frontend IP for squid proxy. A static route is added to LXC containers,
so they could reach the IP.

Change-Id: Ifbb880758e9d83cf1e74070183d028540f43dad1
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>