Ensure that gluster mount is started after service
At the moment there is a race condition possible on
repo container startup, when glusterd.service is
started later then var-www-repo.mount attempts to.

This leads to a mount failure and requires a manual intervention after
each controller restart.

We use BindsTo together with After to ensure that mount will be started
only when glusterd.service
reaches Active state.

A related fix [1] was proposed a while ago, stating that
glusterd.service is started before network-online.target,
which is not a case, as glusterd also starts after network-online.target

[1] https://review.opendev.org/c/openstack/openstack-ansible-plugins/+/845987

Related-Bug: #2104382
Change-Id: I19c95e48eafbecbb8a6bc8bc7b2bf2e7d6b09333
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@advanced.host>