Upgrade setuptools to version 75.8.2
With current version of setuptools `pkg_resources` are not able to
resolve requirements for keystonemiddleware, as they
will fail to find oslo.cache (as it's named as oslo_cache).

This results in pkg_resources.DistributionNotFound exception.

Upgrading setuptools fixes this issue.

This problem is currently a root cause for Magnum to fail loading
Vexxhost CAPI driver.

Change-Id: Id4d39f3c2e0a97a86525ad48da713b5972268550
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@advanced.host>