Ensure locale is always set for Ansible
In some cases, environment from the operator localhost can
be passed to the deploy host, uncluding used locales.

With that, not always the selected locale will be present on the deploy
host, and LC_ALL can be not defined either, for example on Debian 13.

Adding default LC_ALL env var for bootstrap and runtime of
openstack-ansible ensures that locale is
set correcttly. Otherwise, Ansible may fail with
"unsupported locale setting".

Depends-On: https://review.opendev.org/c/openstack/openstack-ansible-plugins/+/965798
Change-Id: I388263019c035744517374fdec2aa5f84408fbf4
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>