Improve nova_console_proxy_types logic
Currently nova_console_proxy_types is defined based on the haproxy host
facts, when it's evaluated as part of haproxy_nova_novnc_console_service

As a result, it will not include all assumed architectures and console
types. Moreover, it will fail to resolve the haproxy service in case
some compute is down due to missing 'architecture' flag.

Proposed patch changes the logic and attempts to check for all available
architectures when concluding nova_console_proxy_types list.
Facts gathering for computes should be accomplished by the nova playbook
as haproxy service configuration is executed as part of it.

Closes-Bug: #2122778
Change-Id: I5a48de05df11c8bff0d82045ec8e1c68bb6c322a
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>