Attempt to fix milestone freeze SHA issue
Whenever I am preparing a milestone bump, in rather
short period of time it becomes unusable and bootstrap is failing to
clone repos from SHAs.

This happens presumably, because local SHA after clone could be replaced
after another merge. We also do a shallow clone, which is a contributing
factor to this.

As an attempt to workaround the issue, we now ask for `git ls-remote`
and use recieved SHA for clone instead.

Change-Id: Ie65431c2a41b2f466edbcf48f2bda27a7fbaeb08
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>