[doc] Refactor contributor documentation
For a while we are having a contradicting sections for contributor
documentation, which makes it unobvious on where to start and what
documentation is relevant.

Patch reduces amount of pages by merging some of repeating
content together.

Distributions page is removed from the contributor
docs, as they  are described in different pages
throughout the documentation.

Change-Id: I716e80ff20b45a6c5df31b994330524a42a33ae8
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>