Don't use nova_console_type for mapping
This causes a circular resolution, as in os_nova role
nova_console_type is calculated based of nova_console_type_mapping.

It was written like that to ensure that defined
nova_console_type will be respected when overriden.

With that change operators need to control
`nova_console_type_mapping` instead of `nova_console_type` to use spice
instead of vnc.

Needed-By: https://review.opendev.org/c/openstack/openstack-ansible-os_nova/+/967052
Change-Id: I25f57b450c0642400a4dce729e0cc66b28cf976e
Signed-off-by: Dmitriy Rabotyagov <dmitriy.rabotyagov@cleura.com>